requirements.txt
discord.py==1.6.0
jsonschema==3.2.0
